% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllLineagesByLoc.R
\name{getAllLineagesByLoc}
\alias{getAllLineagesByLoc}
\title{Retrieve prevalence of lineages by location}
\usage{
getAllLineagesByLoc(
  location,
  other_threshold = 0.05,
  nday_threshold = 10,
  ndays = 180,
  other_exclude,
  cumulative
)
}
\description{
Retrieve prevalence of all lineages over time by location
}
