% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCumulativeBySubadmin.R
\name{getCumulativeBySubadmin}
\alias{getCumulativeBySubadmin}
\title{Retrieve cumulative prevalence of a lineage}
\usage{
getCumulativeBySubadmin(pangolin_lineage, location, mutations, ndays)
}
\description{
Retrieve cumulative prevalence of a PANGO lineage by the immediate administrative level of a location
}
