% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlobalPrevalence.R
\name{getGlobalPrevalence}
\alias{getGlobalPrevalence}
\title{Retrieve global daily prevalence of a lineage}
\usage{
getGlobalPrevalence(pangolin_lineage, mutations = NULL, cumulative = NULL)
}
\description{
Retrieves the global daily prevalence of a PANGO lineage
}
