% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocationData.R
\name{getLocationData}
\alias{getLocationData}
\title{Retrieve COVID-19 data}
\usage{
getLocationData(location_names)
}
\arguments{
\item{location_names:}{vector or list of location (World Bank region, country, state/province, metropolitan area, county) name(s)}
}
\value{
dataframe
}
\description{
Retrieve up-to-date COVID-19 data from outbreak.info for one or more World Bank regions, countries, states/provinces, metropolitan areas, and/or counties.
}
\examples{
getLocationData(c("Texas", "Brazil", "San Diego County"))

}
