% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutAcrossLineage.R
\name{getMutAcrossLineage}
\alias{getMutAcrossLineage}
\title{Retrieve prevalence of mutation across lineages by location}
\usage{
getMutAcrossLineage(mutations, location)
}
\description{
Retrieve prevalence of specified mutation across all lineages by location
}
