% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutDetails.R
\name{getMutDetails}
\alias{getMutDetails}
\title{Retrieve mutation details}
\usage{
getMutDetails(mutations)
}
\arguments{
\item{mutations:}{a `vector` of mutation(s)}
}
\value{
dataframe
}
\description{
Retrieve details of specified mutation(s)
}
\examples{
getMutDetails(mutations=c("S:E484K", "S:N501Y"))

}
