% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalenceByLocation.R
\name{getPrevalenceByLocation}
\alias{getPrevalenceByLocation}
\title{Retrieve daily prevalence of a lineage by location}
\usage{
getPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL
)
}
\description{
Retrieves the daily prevalence of a PANGO lineage by location
}
