% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqCounts.R
\name{getSeqCounts}
\alias{getSeqCounts}
\title{Retrieve sequence count data}
\usage{
getSeqCounts(location = NULL, cumulative = NULL, subadmin = NULL)
}
\description{
Retrieves number of sequences per day by location from outbreak.info
}
