% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubmissionDateByLocation.R
\name{getSubmissionDateByLocation}
\alias{getSubmissionDateByLocation}
\title{Retrieve most recent submission date by location}
\usage{
getSubmissionDateByLocation(
  pangolin_lineage,
  location = NULL,
  mutations = NULL
)
}
\description{
Retrieve most recent submission date by location
}
