% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceByLocation.R
\name{plotPrevalenceByLocation}
\alias{plotPrevalenceByLocation}
\title{Plot daily prevalence of a lineage by location}
\usage{
plotPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL
)
}
\description{
Plots the daily prevalence of a PANGO lineage by location
}
