% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printAPIFields.R
\name{printAPIFields}
\alias{printAPIFields}
\title{Documentation of API fields}
\usage{
printAPIFields()
}
\value{
dataframe
}
\description{
Documents fields included when extracting data from outbreak.info
}
\examples{
View(printAPIFields())

}
