% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllLineagesByLoc.R
\name{getAllLineagesByLoc}
\alias{getAllLineagesByLoc}
\title{Retrieve prevalence of lineages by location}
\usage{
getAllLineagesByLoc(
  location,
  other_threshold = 0.05,
  nday_threshold = 10,
  ndays = 180,
  other_exclude = NULL,
  cumulative = F
)
}
\arguments{
\item{location:}{a location name}

\item{other_threshold:}{minimum prevalence threshold below which lineages are accumulated under "Other" (default=0.05)}

\item{nday_threshold:}{minimum number of days in which the prevalence of a lineage must be below other_threshold to be accumulated under "Other" (default=10)}

\item{ndays:}{the number of days before the current date to be used as a window to accumulate lineages under "Other" (default=180)}

\item{other_exclude:}{(optional) comma separated lineages that are NOT to be included under "Other" even if the conditions specified by the three thresholds above are met}

\item{cumulative:}{`Boolean` (T/F), T returns cumulative prevalence of lineages (default=F)}
}
\value{
dataframe
}
\description{
Retrieve prevalence of all lineages over time by location
}
\examples{
getAllLineagesByLoc(location = "India", other_threshold=0.03, nday_threshold=60, other_exclude="P.1")

}
