% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCollectionDateByLocation.R
\name{getCollectionDateByLocation}
\alias{getCollectionDateByLocation}
\title{Retrieve most recent collection date by location}
\usage{
getCollectionDateByLocation(
  pangolin_lineage,
  location = NULL,
  mutations = NULL
)
}
\description{
Retrieve most recent collection date by location
}
