% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutByLineage.R
\name{getMutByLineage}
\alias{getMutByLineage}
\title{Retrieve mutations by lineage}
\usage{
getMutByLineage(pangolin_lineage, frequency = 0.8)
}
\description{
Retrieve mutations in a specified lineage above an optional frequency threshold
}
