% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrevalenceByLocation.R
\name{plotPrevalenceByLocation}
\alias{plotPrevalenceByLocation}
\title{Plot daily prevalence of a lineage by location}
\usage{
plotPrevalenceByLocation(
  pangolin_lineage,
  location,
  mutations = NULL,
  cumulative = NULL,
  include_title = F
)
}
\arguments{
\item{pangolin_lineage:}{PANGO lineage name}

\item{location:}{a location name}

\item{mutations:}{(optional) a `vector` of mutation(s)}

\item{cumulative:}{(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection}

\item{include_title:}{`Boolean` (T/F), T returns plot with title, F returns plot without title (default=F)}
}
\value{
ggplot2 object
}
\description{
Plots the daily prevalence of a PANGO lineage by location
}
\examples{
plotPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil")


}
